Data collected in an infant eyetracking study

Data from a simple 2-alternative forced choice (2AFC) word recognition task administered to 19- and 24-month-olds. On each trial, infants were shown a picture of an animate object (e.g., a horse) and an inanimate object (e.g., a spoon). After inspecting the images, they disappeared and they heard a label referring to one of them (e.g., "The horse is nearby!"). Finally, the objects re-appeared on the screen and they were prompted to look at the target (e.g., "Look at the horse!").

word_recognition

Format

A data frame with 53940 rows and 10 variables:

ParticipantName

Uniaue participant ID

Sex

M or F

Age

Age, in months

TrialNum

Unique Trial Number

Trial

Name of item shown on trial (also unique for each participant)

TimeFromTrialOnset

Time within trial

Subphase

Subphase within trial (see above)

TimeFromSubphaseOnset

Time within subphase

AOI

Which AOI are they looking at

Animate

Are they looking at the animate AOI?

Inanimate

Are they looking at the inanimate AOI?

TrackLoss

Does current sample not have valid tracking data?

MCDI_Total

Total vocabulary score on MCDI

MCDI_Nouns

Noun vocabulary score on MCDI

MCDI_Verbs

Verb vocabulary score on MCDI

...

Source

Ferguson, B., Graf, E., & Waxman, S. R. (2014). Infants use known verbs to learn novel nouns: Evidence from 15- and 19-month-olds. Cognition, 131(1), 139-146.